requirements.txt
dimorphite_dl==1.3.2
faerun==0.4.7
matplotlib==3.5.3
meeko==0.5.0
mpl_scatter_density==0.7
networkx==2.2
numpy==1.21.5
pandas==1.3.5
python-constraint==1.4.0
pytorch_lightning==1.9.5
rdkit==2023.3.2
rdkit_pypi==2022.9.5
scipy==1.7.3
tmap==1.0.6
torch==1.13.0+cu116
torch_geometric==2.3.1
tqdm==4.64.1
vina==1.2.5
